fixed gallery node
