+++ src/editor.tsx
@@ -3,29 +3,23 @@ import { Value, Editor, Block } from "slate";
 import {
   Editor as SlateReactEditor,
   EventHook,
-  getEventTransfer,
   findDOMNode
 } from "slate-react";
-import AutoReplace from "slate-auto-replace";
+import Html from "slate-html-serializer";
 import { Plugin as SlateReactPlugin } from "slate-react";
-import {
-  PluginConfig,
-  pluginConfigs,
-  EditorButton,
-  PluginsMap
-} from "./plugins";
+import { pluginConfigs } from "./plugins";
 import { StyledMenu, EditorWrapper, StyledContent } from "./editor.css";
-import { mapPropsToComponents } from "./helper/util";
+import { mapPropsToComponents, isEmptyLine, keyMap } from "./helper/util";
 import schemaProps from "./helper/schema";
-import initialValue, { initialEmptyValue } from "./value";
+import initialValue from "./value";
 import { renderNode, renderMark, renderInline } from "./helper/renderer";
 import scrollToCursor from "./helper/scrollToCursor";
-import Html from "slate-html-serializer";
 import { showMenu } from "./helper/showMenu";
 import { getRules } from "./helper/rules";
 import Toolbar from "./components/Toolbar";
 import { Theme } from "./theme.css";
-import snarkdown from "snarkdown";
+import { handlePaste } from "./helper/handlePaste";
+import { LetterpadEditorState, getInitialState } from "./initialState";
 
 // Options available
 export interface LetterpadEditorProps {
@@ -43,23 +37,6 @@ export interface LetterpadEditorProps {
   theme: string;
   width: number;
   hooks(editor: Editor, plugins: object): void;
-}
-
-// Editor state
-interface LetterpadEditorState {
-  menuButtons: EditorButton[];
-  toolbarButtons: EditorButton[];
-  toolbarActive: boolean;
-  toolbarPosition: {
-    top: number;
-    left: number;
-    width: number;
-  };
-  slateReactPlugins: SlateReactPlugin[];
-  pluginsMap: PluginsMap;
-  value: Value;
-  html: string;
-  hooks: object;
 }
 
 /**
@@ -91,8 +68,9 @@ export class LetterpadEditor extends Component<
   // the hover menu
   private menuRef = React.createRef<HTMLDivElement>();
   // handler to serialize/deserialize html
-  private html = new Html({ rules: this.rules });
+  private htmlRules = new Html({ rules: this.rules });
 
+  toolbarPlaceholderStatus = false;
   // default props
   static defaultProps = {
     onButtonClick: () => {},
@@ -112,13 +90,13 @@ export class LetterpadEditor extends Component<
   componentDidMount() {
     this.updateMenu();
 
-    // if there is no html, then lets load some demo data
+    // if there is no html, then load some demo data
     if (this.props.html === null) {
       this.setState({ value: Value.fromJSON(initialValue) });
     } else {
       // If we have received html which is not empty, then deserialize it to create the Document object.
       // This is required by slate editor to render data on page.
-      const value = this.html.deserialize(
+      const value = this.htmlRules.deserialize(
         this.props.html.replace(new RegExp(">[ ]+<", "g"), "><")
       );
       this.setState({ value });
@@ -132,12 +110,6 @@ export class LetterpadEditor extends Component<
     if (prevState.html !== html) {
       this.props.onChange(html);
     }
-
-    setTimeout(() => {
-      if (this.editor) {
-        this.editor.focus();
-      }
-    }, 0);
   };
 
   componentWillUnmount() {
@@ -147,21 +119,19 @@ export class LetterpadEditor extends Component<
   onChange = ({ value }: { value: Value }) => {
     // Whenever there is a change in the editor, we will convert the value to html
     // and save it in the state.
-    const html = this.html.serialize(value);
+    const html = this.htmlRules.serialize(value);
     this.setState({ value, html });
 
-    const { selection, texts, blocks } = value;
+    this.displayLeftToolbarOnNewLine(value);
+    this.hideFloatingMenuBasedOnPlugin(value);
+  };
 
-    if (!value) return;
-    if (!selection) return;
-
-    const isCollapsed = selection.isCollapsed;
-    const topBlock = blocks.get(0);
-    const isAParagraph =
-      topBlock && ["paragraph", "p", "section"].includes(topBlock.type);
-    const isEmptyText = texts && texts.get(0) && texts.get(0).text.length === 0;
-
-    if (isCollapsed && isAParagraph && isEmptyText) {
+  /**
+   * Display the + toolbar when its a new line.
+   */
+  displayLeftToolbarOnNewLine = (value: Value) => {
+    if (isEmptyLine(value) || this.toolbarPlaceholderStatus) {
+      const topBlock = value.blocks.get(0);
       const position = findDOMNode(topBlock).getBoundingClientRect();
       this.setState({
         toolbarActive: true,
@@ -173,19 +143,17 @@ export class LetterpadEditor extends Component<
       });
     } else {
       this.setState({
-        toolbarActive: false,
-        toolbarPosition: {
-          left: this.state.toolbarPosition.left,
-          top: this.state.toolbarPosition.top,
-          width: 0
-        }
+        toolbarActive: false
       });
     }
-    /**
-     * Few plugins of `block` type might not allow further transformations from other plugins.
-     * eg. Codeblocks will not allow any other transformation (like bold, italic)
-     * to work inside its block.
-     */
+  };
+
+  /**
+   * Few plugins of `block` type might not allow further transformations from other plugins.
+   * eg. Codeblocks will not allow any other transformation (like bold, italic)
+   * to work inside its block.
+   */
+  hideFloatingMenuBasedOnPlugin = (value: Value) => {
     let parentPlugin = null;
     // for blocks (h1, blockquote,codeblock, etc)
     if (value.fragment.nodes.size > 0) {
@@ -210,35 +178,12 @@ export class LetterpadEditor extends Component<
 
   onPaste: EventHook<React.ClipboardEvent> = (event, editor) => {
     scrollToCursor();
-
-    const transfer = getEventTransfer(event);
-    if (transfer.type != "html") {
-      // convert markdown to html
-      let html = snarkdown((transfer as any).text);
-      const { document } = this.html.deserialize(html);
-      return editor.insertFragment(document);
-    }
-
-    const parentTag = editor.value.blocks.first().type; // p, pre, etc
-    for (let i = 0; i < pluginConfigs.length; i++) {
-      const config = pluginConfigs[i];
-      if (config.onPasteReturnHtml === false) {
-        if (config.identifier && config.identifier.indexOf(parentTag) >= 0) {
-          return editor.insertText((transfer as any).text);
-        }
-      }
-    }
-    // remove style attr
-    const REMOVE_STYLE_ATTR = /style="[^\"]*"/gi;
-    let html = (transfer as any).html.replace(REMOVE_STYLE_ATTR, "");
-    // TODO: fix the transfer as any
-    const { document } = this.html.deserialize(html);
-    editor.insertFragment(document);
+    handlePaste(event, editor, this.htmlRules);
   };
 
   hideMenu = (e: KeyboardEvent) => {
     const { current } = this.menuRef;
-    if (current != null && e.keyCode === 27) {
+    if (current != null && e.keyCode === keyMap.ESCAPE) {
       current.removeAttribute("style");
     }
   };
@@ -252,15 +197,14 @@ export class LetterpadEditor extends Component<
   };
 
   triggerWordCountHook = (value: Value) => {
-    const totalBlocks = value.document.getBlocks();
-    if (totalBlocks.size > 0) {
-      let charCount = 0;
-      charCount = totalBlocks.reduce((memo: number | undefined, b: any) => {
-        return memo + b.text.trim().split(/\s+/).length;
-      }, 0);
-      if (this.props.getCharCount) {
-        this.props.getCharCount(charCount);
-      }
+    let wordCount = 0;
+
+    for (const [node] of value.document.blocks({ onlyLeaves: true })) {
+      const words = node.text.trim().split(/\s+/);
+      wordCount += words.length;
+    }
+    if (this.props.getCharCount) {
+      this.props.getCharCount(wordCount);
     }
   };
 
@@ -300,6 +244,9 @@ export class LetterpadEditor extends Component<
           position={this.state.toolbarPosition}
           editor={this.editor}
           buttons={this.state.toolbarButtons}
+          setPlaceholderStatus={(status: boolean) =>
+            (this.toolbarPlaceholderStatus = status)
+          }
           data={data}
         />
       </>
@@ -356,75 +303,6 @@ export class LetterpadEditor extends Component<
       </Theme>
     );
   }
-}
-
-/**
- * Initial state for the Letterpad Editor
- */
-function getInitialState(pluginConfigs: PluginConfig[]): LetterpadEditorState {
-  const menuButtons: LetterpadEditorState["menuButtons"] = [];
-  const toolbarButtons: LetterpadEditorState["toolbarButtons"] = [];
-  const slateReactPlugins: LetterpadEditorState["slateReactPlugins"] = [];
-  const pluginsMap: LetterpadEditorState["pluginsMap"] = {
-    node: {},
-    mark: {},
-    inline: {}
-  };
-  const hooks = {};
-
-  // Collect all necessary things
-  for (const pluginConfig of pluginConfigs) {
-    // collect menubuttons
-    if (pluginConfig.menuButtons != null) {
-      menuButtons.push(...pluginConfig.menuButtons);
-    }
-
-    // collect toolbar buttons
-    if (pluginConfig.toolbarButtons != null) {
-      toolbarButtons.push(...pluginConfig.toolbarButtons);
-    }
-
-    // collect slate react plugins
-    if (pluginConfig.slatePlugin != null) {
-      slateReactPlugins.push(pluginConfig.slatePlugin());
-    }
-
-    // collect slate react plugins from markdown config
-    if (pluginConfig.markdown != null) {
-      slateReactPlugins.push(AutoReplace(pluginConfig.markdown));
-    }
-
-    // collect all the plugin utility functions
-    if (pluginConfig.hooks != null) {
-      (hooks as any)[pluginConfig.name] = { ...pluginConfig.hooks };
-    }
-
-    let { identifier, renderType } = pluginConfig;
-    if (identifier != null && renderType != null) {
-      identifier.forEach(id => {
-        pluginsMap[renderType as keyof PluginsMap][id] = {
-          plugin: pluginConfig,
-          is: id
-        };
-      });
-    }
-  }
-
-  return {
-    menuButtons,
-    toolbarButtons,
-    slateReactPlugins,
-    pluginsMap,
-    value: Value.fromJSON(initialEmptyValue),
-    html: "",
-    toolbarActive: false,
-    hooks,
-    toolbarPosition: {
-      top: 0,
-      left: 0,
-      width: 0
-    }
-  };
 }
 
 export default LetterpadEditor;
